refactor(database): split storage utils into indexedDB and localStorage classes
- Create separate IndexedDB class for database operations (conversations, messages, presets)
- Create separate LocalStorage class for config/theme operations
- Update all imports and usages across components and contexts
- Move migration logic to dedicated migration.ts file
- Convert StorageUtils to static class methods for better organization

This refactoring improves code organization by separating concerns between
IndexedDB and LocalStorage storage mechanisms.
